pyproject.toml
[project]
name = "banbu-home-agent"
version = "0.1.0"
description = "Scene-driven smart-home agent with ContextPilot + local llama.cpp"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "openai>=1.30",
    "contextpilot",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
banbu-probe = "banbu.cli.probe:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["banbu*"]
